+++ PCIT_core/internal/llvm_interface/include/class_impls/native_ptr_decls.h
@@ -20,6 +20,8 @@ namespace llvm{
 	class NoFolder;
 	class IRBuilderDefaultInserter;
 
+	class ExecutionEngine;
+	
 	class Function;
 
 	// types