+++ PCIT_core/internal/llvm_interface/include/llvm_interface.h
@@ -19,4 +19,5 @@
 #include "./Function.h"
 
 #include "./IRBuilder.h"
+#include "./ExecutionEngine.h"